namada_parameters
Module storage

Parameters storage

Structs

Key

A storage key is made of storage key segments DbKeySeg, separated by KEY_SEGMENT_SEPARATOR.

Functions

get_active_status_key_at_addr

get_bridge_contract_address_key_at_addr

get_epoch_duration_key_at_addr

get_epoch_duration_storage_key

Storage key used for epoch parameter.

get_epochs_per_year_key

Storage key used for epochs_per_year parameter.

get_epochs_per_year_key_at_addr

get_eth_start_height_key_at_addr

get_gas_cost_key

Storage key used for the gas cost table

get_gas_scale

Helper function to retrieve the gas_scale protocol parameter from storage

get_gas_scale_key

Storage key used for the gas scale

get_gas_scale_key_at_addr

get_implicit_vp_key

Storage key used for implicit VP parameter.

get_implicit_vp_key_at_addr

get_masp_epoch_multiplier_key

Storage key used for masp_epoch_multiplier parameter.

get_masp_epoch_multiplier_key_at_addr

get_masp_fee_payment_gas_limit_key

Storage key used for the fee unshielding gas limit

get_masp_fee_payment_gas_limit_key_at_addr

get_max_block_gas

Helper function to retrieve the max_block_gas protocol parameter from storage

get_max_block_gas_key

Storage key used for the max block gas.

get_max_block_gas_key_at_addr

get_max_proposal_bytes_key

Storage key used for the max proposal bytes.

get_max_proposal_bytes_key_at_addr

get_max_tx_bytes_key

Storage key used for the max tx bytes.

get_max_tx_bytes_key_at_addr

get_min_confirmations_key_at_addr

get_minimum_gas_price_key_at_addr

get_native_erc20_key_at_addr

get_native_token_transferable_key

Storage key used for the flag to enable the native token transfer

get_native_token_transferable_key_at_addr

get_tx_allowlist_key_at_addr

get_tx_allowlist_storage_key

Storage key used for tx allowlist parameter.

get_vp_allowlist_key_at_addr

get_vp_allowlist_storage_key

Storage key used for vp allowlist parameter.

is_active_status_key_at_addr

is_bridge_contract_address_key_at_addr

is_epoch_duration_key_at_addr

is_epoch_duration_storage_key

Returns if the key is an epoch storage key.

is_epochs_per_year_key

Returns if the key is the epoch_per_year key.

is_epochs_per_year_key_at_addr

is_eth_start_height_key_at_addr

is_gas_scale_key_at_addr

is_implicit_vp_key

Returns if the key is the implicit VP key.

is_implicit_vp_key_at_addr

is_masp_epoch_multiplier_key_at_addr

is_masp_fee_payment_gas_limit_key_at_addr

is_max_block_gas_key_at_addr

is_max_proposal_bytes_key

Returns if the key is the max proposal bytes key.

is_max_proposal_bytes_key_at_addr

is_max_tx_bytes_key

Returns if the key is the max tx bytes key.

is_max_tx_bytes_key_at_addr

is_min_confirmations_key_at_addr

is_minimum_gas_price_key_at_addr

is_native_erc20_key_at_addr

is_native_token_transferable

Helper function to retrieve the is_native_token_transferable protocol parameter from storage

is_native_token_transferable_key_at_addr

is_parameter_key

Returns if the key is a parameter key.

is_protocol_parameter_key

Returns if the key is a protocol parameter key.

is_tx_allowlist_key

Returns if the key is the tx_allowlist key.

is_tx_allowlist_key_at_addr

is_vp_allowlist_key

Returns if the key is the vp_allowlist key.

is_vp_allowlist_key_at_addr